flask backend works
when the button is clicked, javascript makes a call to the flask backend, which attempts to read a document named 'report.xlsx' in the project directory, which means the user must place that file there before using. This only works in the specific format the report is generated by b2w as of today. If the report format changes in the future, this will break the program.